+++ 07_Raycasting/Material.h
@@ -9,17 +9,17 @@ private:
 	Vec3 diffuse;
 	Vec3 specular;
 
-	double exponent;
-	double local;
-	std::optional<double> IOR;
+	float exponent;
+	float local;
+	std::optional<float> IOR;
 	bool m_isShadowCaster;
 
 public:
-	Material(const Vec3& ambient, const Vec3& diffuse, const Vec3& specular, double exponent)
+	Material(const Vec3& ambient, const Vec3& diffuse, const Vec3& specular, float exponent)
 		: Material(ambient, diffuse, specular, exponent, 1.0)
 	{ }
 
-	Material(const Vec3& ambient, const Vec3& diffuse, const Vec3& specular, double exponent, double local)
+	Material(const Vec3& ambient, const Vec3& diffuse, const Vec3& specular, float exponent, float local)
 		: Material(ambient, diffuse, specular, exponent, local, std::nullopt)
 	{ }
 	
@@ -31,11 +31,11 @@ public:
 	///						that gets reflected or reflected and refracted according to Schlick's approximation of the
 	///						Fresnel equations</param>
 	/// <param name="IOR">Index Of Refraction</param>
-	Material(const Vec3& ambient, const Vec3& diffuse, const Vec3& specular, double exponent, double local, const std::optional<double>& IOR)
+	Material(const Vec3& ambient, const Vec3& diffuse, const Vec3& specular, float exponent, float local, const std::optional<float>& IOR)
 		: Material(ambient, diffuse, specular, exponent, local, IOR, true)
 	{ }
 
-	Material(const Vec3& ambient, const Vec3& diffuse, const Vec3& specular, double exponent, double local, const std::optional<double>& IOR, bool isShadowCaster)
+	Material(const Vec3& ambient, const Vec3& diffuse, const Vec3& specular, float exponent, float local, const std::optional<float>& IOR, bool isShadowCaster)
 		: ambient(ambient), diffuse(diffuse), specular(specular), exponent(exponent), local(local), IOR(IOR), m_isShadowCaster(isShadowCaster)
 	{ }
 	
@@ -45,7 +45,7 @@ public:
 
 	Vec3 getSpecular() const;
 
-	double getExp() const;
+	float getExp() const;
 
 	bool isShadowCaster() const;
