+++ 07_Raycasting/Sphere.h
@@ -5,7 +5,7 @@ class Sphere : public IntersectableObject
 {
 
 public:
-	Sphere(const Vec3& center, double radius, const Material& material);
+	Sphere(const Vec3& center, float radius, const Material& material);
     virtual ~Sphere() {}
     
 	Material getMaterial() const override;
