+++ 08_Raytracing/Intersection.h
@@ -7,16 +7,16 @@ class Intersection
 private:
 	Material material;
 	Vec3 normal;
-	double t;
+	float t;
 
 public:
-	Intersection(const Material& material, const Vec3& normal, double t)
+	Intersection(const Material& material, const Vec3& normal, float t)
 		: material(material), normal(normal), t(t)
 	{ };
 
 	Material getMaterial() const;
 	Vec3 getNormal() const;
-	double getT() const;
+	float getT() const;
 
 };
