+++ 08_Raytracing/Sphere.cpp
@@ -1,6 +1,6 @@
 #include "Sphere.h"
 
-Sphere::Sphere(const Vec3& center, double radius, const Material& material)
+Sphere::Sphere(const Vec3& center, float radius, const Material& material)
 	: center(center), sqradius(radius* radius), material(material)
 {
 
@@ -15,16 +15,16 @@ std::optional<Intersection> Sphere::intersect(const Ray& ray) const
 {
 	Vec3 l = center - ray.getOrigin();
 
-	double tCenter = Vec3::dot(l, ray.getDirection());
+	float tCenter = Vec3::dot(l, ray.getDirection());
 	if (tCenter < 0)
 		return {};	// no intersection
 
-	double dSq = l.sqlength() - tCenter * tCenter;
+	float dSq = l.sqlength() - tCenter * tCenter;
 	if (dSq > sqradius)
 		return {};	// no intersection
 
-	double dist = sqrt(sqradius - dSq);
-	double t = tCenter - dist;
+	float dist = sqrt(sqradius - dSq);
+	float t = tCenter - dist;
 
 	if (t < 0)
 		t = tCenter + dist;	// when inside sphere
